+++ SplunkOpenTelemetry/Sources/SplunkOpenTelemetry/Log Event Processors/OTLPLogToSpanEventProcessor.swift
@@ -17,7 +17,6 @@ limitations under the License.
 
 import Foundation
 import OpenTelemetryApi
-import OpenTelemetryProtocolExporterCommon
 import OpenTelemetrySdk
 import SplunkCommon
 import SplunkOpenTelemetryBackgroundExporter