+++ SplunkOpenTelemetry/Sources/SplunkOpenTelemetry/Trace Processor/OTLPTraceProcessor.swift
@@ -17,7 +17,6 @@ limitations under the License.
 
 import Foundation
 import OpenTelemetryApi
-import OpenTelemetryProtocolExporterCommon
 import OpenTelemetrySdk
 import SplunkCommon
 import SplunkOpenTelemetryBackgroundExporter
@@ -48,7 +47,7 @@ public class OTLPTraceProcessor: TraceProcessor {
         accessToken: String? = nil
     ) {
 
-        let configuration = OtlpConfiguration()
+        let configuration = OTLPExporterConfiguration()
         let envVarHeaders: [(String, String)] = []
         var headers: [String: String] = [:]