+++ SplunkOpenTelemetryBackgroundExporter/Sources/SplunkOpenTelemetryBackgroundExporter/OTLPBackgroundHTTPBaseExporter.swift
@@ -1,6 +1,6 @@
 //
 /*
-Copyright 2025 Splunk Inc.
+Copyright 2026 Splunk Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -18,8 +18,6 @@ limitations under the License.
 import CiscoDiskStorage
 import CiscoEncryption
 import Foundation
-import OpenTelemetryProtocolExporterCommon
-import SwiftProtobuf
 
 /// Basic implementation of exporters.
 public class OTLPBackgroundHTTPBaseExporter {
@@ -36,7 +34,7 @@ public class OTLPBackgroundHTTPBaseExporter {
     var endpoint: URL?
     let envVarHeaders: [(String, String)]?
     var additionalHeaders: [String: String]
-    let config: OtlpConfiguration
+    let config: OTLPExporterConfiguration
     let diskStorage: DiskStorage
     var checkStalledTask: Task<Void, Never>?
 
@@ -59,9 +57,9 @@ public class OTLPBackgroundHTTPBaseExporter {
 
     public init(
         endpoint: URL?,
-        config: OtlpConfiguration = OtlpConfiguration(),
+        config: OTLPExporterConfiguration = OTLPExporterConfiguration(),
         qosConfig: SessionQOSConfiguration,
-        envVarHeaders: [(String, String)]? = EnvVarHeaders.attributes,
+        envVarHeaders: [(String, String)]? = OTLPEnvVarHeaders.attributes,
         headers: [String: String] = [:],
         diskStorage: DiskStorage = FilesystemDiskStorage(
             prefix: FilesystemPrefix(module: "OTLPBackgroundExporter"),